requirements.txt
flask
pillow
pytesseract
opencv-python
python-dotenv
google-generativeai
numpy
